chore: Refactor handle_info function in chat_server_listener.erl
The commit refactors the handle_info function in the chat_server_listener.erl module. It adds logic to handle incoming TCP connections and validate usernames. If the username is valid, it spawns a user handler process and sends a welcome message. If the username is invalid, it sends an error message and closes the connection.